feat: report profile
