+++ apps/web/src/components/Watch/Comments/CommentOptions.tsx
@@ -1,7 +1,7 @@
 import FlagOutline from '@components/Common/Icons/FlagOutline'
 import ThreeDotsOutline from '@components/Common/Icons/ThreeDotsOutline'
 import TrashOutline from '@components/Common/Icons/TrashOutline'
-import ReportPublication from '@components/ReportPublication'
+import ReportPublication from '@components/Report/Publication'
 import Confirm from '@components/UIElements/Confirm'
 import useHandleWrongNetwork from '@hooks/useHandleWrongNetwork'
 import useProfileStore from '@lib/store/profile'
